+++ code/math/逆元组合数.cpp
@@ -4,16 +4,16 @@ int facinv[N], fac[N];
 void init_fac() {
     fac[0] = fac[1] = 1;
     for(int i = 2; i < N; i++) {
-        fac[i] = fac[i - 1] * i % mod;
+        fac[i] = (ll) fac[i - 1] * i % mod;
     }
     facinv[N - 1] = ksm(fac[N - 1], mod - 2);
     for(int i = N - 2; i >= 0; i--) {
-    	facinv[i] = facinv[i + 1] * (i + 1) % mod;
+    	facinv[i] = (ll) facinv[i + 1] * (i + 1) % mod;
 	}
 }
 int binom(int n, int k) {
     if(n < 0 || k < 0 || k > n) { return 0; }
-    return fac[n] * facinv[n - k] % mod * facinv[k] % mod;
+    return (ll) fac[n] * facinv[n - k] % mod * facinv[k] % mod;
 }
 
 int inv[N];
